fix(lint): cria utilitário dedicado para travamento de rolagem compartilhado
- Cria `frontend/src/utils/scrollLock.ts` com as funções `lockScroll` e `unlockScroll` extraídas de `Modal.tsx`
- Remove a lógica de contagem de modais ativos de dentro do componente `Modal`
- Atualiza os imports de `Drawer.tsx` e `ProductImageGallery.tsx` para o novo caminho `@/utils/scrollLock`
- Corrige o estado de carregamento do iframe no checkout com cartão, ativando-o apenas no momento da submissão do pagamento
